# backend/src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cinemate API",
    "description": "Full-stack movie streaming and social recommendation platform. Catalog data from PostgreSQL; rich movie detail, cast, and trailers fetched live from TMDB API.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/api", "description": "Reverse-proxied through Nginx" }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly JWT cookie set on login/register. Access token TTL 15 min; refresh token 7 days."
      }
    },
...
  },
  "paths": {
...
    "/recommendations": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Personalised movie feed — 4-tier cold-start: popularity → ALS → content-based → hybrid CF+CB",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ movies[], strategy: 'popular'|'als'|'content'|'hybrid' }" } }
      }
    },
    "/recommendations/similar/{tmdbId}": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Genome-based similar movies (default 6)",
        "parameters": [{ "name": "tmdbId", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "{ movies[] }" } }
      }
    },
...
    "/recommendations/mood": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Genome mood-based discovery",
        "parameters": [
          { "name": "moods",  "in": "query", "schema": { "type": "string" }, "description": "Comma-separated mood names" },
          { "name": "genre",  "in": "query", "schema": { "type": "string" } },
          { "name": "limit",  "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "{ movies[] }" } }
      }
    },
    "/recommendations/mood-context": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Advanced filtered discovery: mood + duration + context + multi-genre + exclusions; optional CF blend when authenticated",
        "parameters": [
          { "name": "moods",        "in": "query", "schema": { "type": "string" } },
          { "name": "duration",     "in": "query", "schema": { "type": "string", "enum": ["short","medium","long","any"] } },
          { "name": "context",      "in": "query", "schema": { "type": "string", "enum": ["solo","friends","date","family","any"] } },
          { "name": "genre",        "in": "query", "schema": { "type": "string" }, "description": "Comma-separated genre names (ANY match)" },
          { "name": "exclude_moods","in": "query", "schema": { "type": "string" } },
          { "name": "limit",        "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "{ movies[], strategy }" } }
      }
    },
    "/recommendations/natural-search": {
      "post": {
        "tags": ["Recommendations"],
        "summary": "GPT-4o-mini natural language → structured recommendation params (requires OPENAI_API_KEY; 503 if absent)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["query"], "properties": { "query": { "type": "string", "example": "something scary to watch alone at night" } } } } } },
        "responses": {
          "200": { "description": "{ movies[], parsed: { moods, genres, duration, context, exclude_moods, similar_to, explanation }, strategy }" },
          "503": { "description": "OPENAI_API_KEY not configured" }
        }
      }
    },
    "/recommendations/user-genome": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "Genome tag profile for current user — top 25 tags from liked movies (score ≥ 7) + top 15 distinctive tags",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ tags: [{ tag, score, distinctive }], n_liked }" } }
      }
    },
...
    "/recommendations/refresh": {
      "post": {
        "tags": ["Recommendations"],
        "summary": "Retrain ML models from current DB data (admin only)",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ status: 'refreshed', movies, ratings, elapsed_s }" } }
      }
    },
    "/recommendations/evaluate": {
      "get": {
        "tags": ["Recommendations"],
        "summary": "RMSE + MAE on 20% held-out split (admin only)",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ rmse, mae, n_test }" } }
      }
    },
...
    "/admin/insights/top-movies": {
      "get": {
        "tags": ["Admin"],
        "summary": "Top 20 most-watched films platform-wide with avg completion %",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ movies: [{ title, poster_path, watch_count, avg_completion_pct }] }" } }
      }
    },
    "/admin/insights/abandonment": {
      "get": {
        "tags": ["Admin"],
        "summary": "Films with highest abandonment rate (<20% completion, ≥3 watchers)",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ movies[] }" } }
      }
    },
    "/admin/insights/active-users": {
      "get": {
        "tags": ["Admin"],
        "summary": "DAU (24h), WAU (7d), MAU (30d) based on watch_history",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ dau, wau, mau }" } }
      }
    },
    "/admin/insights/recommendation-ctr": {
      "get": {
        "tags": ["Admin"],
        "summary": "Platform-wide recommendation acceptance rate by strategy",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ by_strategy: [{ strategy, shown, watched, rate }] }" } }
      }
    },
    "/admin/insights/genre-breakdown": {
      "get": {
        "tags": ["Admin"],
        "summary": "Total watch-hours and unique viewers by genre",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "{ genres: [{ name, total_hours, unique_viewers }] }" } }
      }
    }
  }
}